backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
httpx==0.28.1
openai>=1.0.0
python-dotenv==1.0.0
pydantic>=2.0.0
python-multipart
supabase==2.27.0
websockets>=15.0
